decoupled Searchbar component from Dropdown and put api states in a ApiContext file
