+++ client/src/components/Home_components/Home.jsx
@@ -1,10 +1,17 @@
 import React from 'react';
 import Dropdown from './Dropdown';
+import Searchbar from './Searchbar'
+import {ApiProvider} from '../context/ApiContext'
 
 const Home = () => {
   return (
     <div className="home">
-      <Dropdown />
+      <div className="center-container">
+        <ApiProvider>
+          <Dropdown />
+          <Searchbar />
+        </ApiProvider>
+      </div>
     </div>
   );
 };
